[refactor] DiaryWritingViewController 코드를 별도의 extension으로 책임 분리
